+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeClient.kt
@@ -13,6 +13,7 @@ import com.braintreepayments.api.IntegrationType.Integration
  * Core Braintree class that handles network requests.
  */
 @Suppress("LargeClass", "LongParameterList", "TooManyFunctions")
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 open class BraintreeClient @VisibleForTesting internal constructor(
 
     /**
@@ -76,6 +77,7 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      * @param authorization The tokenization key or client token to use. If an invalid authorization
      * is provided, a [BraintreeException] will be returned via callback.
      */
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     constructor(context: Context, authorization: String) :
             this(BraintreeOptions(context = context, initialAuthString = authorization))
 
@@ -86,6 +88,7 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      * @param clientTokenProvider An implementation of [ClientTokenProvider] that [BraintreeClient]
      * will use to fetch a client token on demand.
      */
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     constructor(context: Context, clientTokenProvider: ClientTokenProvider) :
             this(BraintreeOptions(context = context, clientTokenProvider = clientTokenProvider))
 
@@ -104,7 +107,8 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      * authorization is provided, a [BraintreeException] will be returned via callback.
      * @param returnUrlScheme A custom return url to use for browser and app switching
      */
-    constructor (context: Context, authorization: String, returnUrlScheme: String) : this(
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+    constructor (context: Context, authorization: String, returnUrlScheme: String?) : this(
         BraintreeOptions(
             context = context,
             initialAuthString = authorization,
@@ -127,6 +131,7 @@ open class BraintreeClient @VisibleForTesting internal constructor(
      * will use to fetch a client token on demand.
      * @param returnUrlScheme     A custom return url to use for browser and app switching
      */
+    @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     constructor(
         context: Context,
         clientTokenProvider: ClientTokenProvider,