+++ Card/src/main/java/com/braintreepayments/api/CardClient.java
@@ -1,5 +1,7 @@
 package com.braintreepayments.api;
 
+import android.content.Context;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.annotation.VisibleForTesting;
@@ -17,7 +19,18 @@ public class CardClient {
     private final BraintreeClient braintreeClient;
     private final ApiClient apiClient;
 
-    public CardClient(@NonNull BraintreeClient braintreeClient) {
+    /**
+     * Initializes a new {@link CardClient} instance
+     *
+     * @param context an Android Context
+     * @param authorization a Tokenization Key or Client Token used to authenticate
+     */
+    public CardClient(@NonNull Context context, @NonNull String authorization) {
+        this(new BraintreeClient(context, authorization));
+    }
+
+    @VisibleForTesting
+    CardClient(@NonNull BraintreeClient braintreeClient) {
         this(braintreeClient, new ApiClient(braintreeClient));
     }