+++ Demo/src/main/java/com/braintreepayments/demo/BaseFragment.java
@@ -6,8 +6,9 @@ import androidx.annotation.CallSuper;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentActivity;
 
-import com.braintreepayments.api.BraintreeClient;
 import com.braintreepayments.api.PaymentMethodNonce;
+
+import java.util.Objects;
 
 public abstract class BaseFragment extends Fragment {
 
@@ -47,11 +48,15 @@ public abstract class BaseFragment extends Fragment {
         }
     }
 
-    protected BraintreeClient getBraintreeClient() {
+
+    protected void fetchAuthorization(BraintreeAuthorizationCallback callback) {
         DemoActivity demoActivity = getDemoActivity();
         if (demoActivity != null) {
-            return demoActivity.getBraintreeClient();
+            demoActivity.fetchAuthorization(callback);
         }
-        return null;
+    }
+
+    String getAuthStringArg() {
+        return Objects.requireNonNull(requireArguments().getString("authString"));
     }
 }