+++ Demo/src/main/java/com/braintreepayments/demo/SEPADirectDebitFragment.java
@@ -9,7 +9,6 @@ import android.widget.Button;
 import androidx.annotation.NonNull;
 import androidx.navigation.fragment.NavHostFragment;
 
-import com.braintreepayments.api.BraintreeClient;
 import com.braintreepayments.api.PostalAddress;
 import com.braintreepayments.api.SEPADirectDebitClient;
 import com.braintreepayments.api.SEPADirectDebitLauncher;
@@ -30,8 +29,7 @@ public class SEPADirectDebitFragment extends BaseFragment {
         Button button = view.findViewById(R.id.sepa_direct_debit_button);
         button.setOnClickListener(this::launchSEPADirectDebit);
 
-        BraintreeClient braintreeClient = getBraintreeClient();
-        sepaDirectDebitClient = new SEPADirectDebitClient(braintreeClient);
+        sepaDirectDebitClient = new SEPADirectDebitClient(requireContext(), super.getAuthStringArg());
 
         sepaDirectDebitLauncher = new SEPADirectDebitLauncher(sepaDirectDebitBrowserSwitchResult ->
             sepaDirectDebitClient.tokenize(sepaDirectDebitBrowserSwitchResult, (sepaDirectDebitNonce, error) -> {